requirements.txt
mkdocs-material[imaging]
mkdocs-rss-plugin
mkdocs-macros-plugin
Jinja2
ruamel.yaml
faqtory
